test_create_post_success, test_delete_comment_success 두 개 빼고 test 완료

+++ app/service/comment_service.py
@@ -34,8 +34,9 @@ class CommentService:
 
     def delete(self, comment_id: int):
         comment = self.db.query(Comment).filter(Comment.id == comment_id).first()
-        self.db.delete(comment)
-        self.db.commit()
+        if comment:
+            self.db.delete(comment)
+            self.db.commit()  # 반드시 커밋 호출
 
     def _get_comment_by_id(self, comment_id: int) -> Comment:
         comment = self.db.query(Comment).filter(Comment.id == comment_id).first()
